+++ ElvUI_LuckyoneUI/Core/Update.lua
@@ -5,9 +5,10 @@ function L1UI:UpdateLayout(layout)
 	-- Bump install version
 	E.db.L1UI.install_version = L1UI.Version
 
-	-- WindTools updates
-	if E:IsAddOnEnabled('ElvUI_WindTools') and E.Retail then
-		L1UI:Setup_WindTools(true)
+	if E.Retail then
+		-- General db updates
+		E.db.general.altPowerBar.statusBarColorGradient = true
+		E.db.general.itemLevel.itemLevelFontSize = 10
 	end
 
 	L1UI:Print(L["Successfully updated your current layout to LuckyoneUI version "] .. L1UI.Version)